django/requirements.txt
Django==3.2.5
uwsgi==2.0.20
django-environ==0.4.5
channels
channels_redis
djangorestframework
django-cors-headers~=3.13.0
dj-rest-auth
mysqlclient